+++ RPG.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPG", "RPGTest\RPG.csproj", "{BE6C7FE4-E607-4205-8E00-377341791CF3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPG", "RPG\RPG.csproj", "{BE6C7FE4-E607-4205-8E00-377341791CF3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPGTests", "UnitTestTests\RPGTests.csproj", "{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RPGTests", "RPGTests\RPGTests.csproj", "{B492D835-6CDF-4AB9-A996-337A42D259AF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,14 +21,14 @@ Global
 		{BE6C7FE4-E607-4205-8E00-377341791CF3}.Release|Any CPU.ActiveCfg = Release|x86
 		{BE6C7FE4-E607-4205-8E00-377341791CF3}.Release|x86.ActiveCfg = Release|x86
 		{BE6C7FE4-E607-4205-8E00-377341791CF3}.Release|x86.Build.0 = Release|x86
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Debug|x86.Build.0 = Debug|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Release|x86.ActiveCfg = Release|Any CPU
-		{0BDECAD7-DF26-4AEB-ADD8-7B3CE956248E}.Release|x86.Build.0 = Release|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Debug|x86.Build.0 = Debug|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Release|x86.ActiveCfg = Release|Any CPU
+		{B492D835-6CDF-4AB9-A996-337A42D259AF}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
