+++ RPG/Characters/GraphicCharacter.cs
