+++ RPG/Properties/AssemblyInfo.cs
