+++ RPG/Skills/Burn.cs
