+++ RPG/Skills/CritDamageEffect.cs
