+++ RPG/Skills/IEffect.cs
