+++ RPG/Skills/StatusEffects/Bleeding.cs
@@ -6,26 +6,26 @@ using System.Text;
 
 namespace RPG.Skills.StatusEffects
 {
-    public class Burning : IStatuseffect
+    public class Bleeding : IStatuseffect
     {
         public int Duration { get; set; }
         public int Damage { get; set; }
 
-        public Burning(Character source)
+        public Bleeding(Character source)
         {
             Random r1 = new Random();
-            
-            this.Duration = Convert.ToInt32(r1.Next(1, 4 * 1000) / 1000);
-            this.Damage = Convert.ToInt32((r1.Next(source.FightMagic / 5, (source.FightMagic / 3) * 1000)) / 1000);
+
+            Duration = Convert.ToInt32(r1.Next(3, 6 * 1000) / 1000);
+            Damage = Convert.ToInt32((r1.Next(source.FightMagic / 7, (source.FightMagic / 5) * 1000)) / 1000);
         }
 
-        public int ExecuteStatus(Character target)
+        public int ExecuteStatus()
         {
             Duration--;
             return Damage;
         }
 
-        public bool IsDone ()
+        public bool IsDone()
         {
             if (Duration <= 0)
                 return true;
