+++ RPG/Skills/StatusEffects/Burning.cs
@@ -6,28 +6,26 @@ using System.Text;
 
 namespace RPG.Skills.StatusEffects
 {
-    public class Mindblown : IStatuseffect
+    public class Burning : IStatuseffect
     {
         public int Duration { get; set; }
         public int Damage { get; set; }
 
-        public Mindblown(Character source, Character target)
+        public Burning(Character source)
         {
-            Damage = 0;
-          
-            if (target.Level > source.Level)
-                Duration = 1;
-            else
-                Duration = source.Level - target.Level;
+            Random r1 = new Random();
+            
+            this.Duration = Convert.ToInt32(r1.Next(1, 4 * 1000) / 1000);
+            this.Damage = Convert.ToInt32((r1.Next(source.FightMagic / 5, (source.FightMagic / 3) * 1000)) / 1000);
         }
 
-        public int ExecuteStatus(Character target)
+        public int ExecuteStatus()
         {
             Duration--;
             return Damage;
         }
 
-        public bool IsDone()
+        public bool IsDone ()
         {
             if (Duration <= 0)
                 return true;
