+++ RPG/Skills/StatusEffects/IStatuseffect.cs
